Fix rpc
