+++ substrate/frame/revive/rpc/src/example.rs
@@ -157,6 +157,7 @@ impl<Client: EthRpcClient + Send + Sync> TransactionBuilder<Client> {
 			.await
 			.with_context(|| "Failed to fetch gas estimate")?;
 
+		println!("Gas estimate: {gas:?}");
 		let mut unsigned_tx = TransactionLegacyUnsigned {
 			gas,
 			nonce,
