+++ substrate/frame/revive/rpc/src/receipt_provider/cache.rs
@@ -16,7 +16,7 @@
 // limitations under the License.
 use super::ReceiptProvider;
 use jsonrpsee::core::async_trait;
-use pallet_revive::evm::{ReceiptInfo, TransactionSigned, H256, U256};
+use pallet_revive::evm::{ReceiptInfo, TransactionSigned, H256};
 use std::{collections::HashMap, sync::Arc};
 use tokio::sync::RwLock;
 
@@ -48,13 +48,13 @@ impl ReceiptProvider for CacheReceiptProvider {
 	async fn receipt_by_block_hash_and_index(
 		&self,
 		block_hash: &H256,
-		transaction_index: &U256,
+		transaction_index: usize,
 	) -> Option<ReceiptInfo> {
 		let cache = self.cache().await;
 		let receipt_hash = cache
 			.transaction_hashes_by_block_and_index
 			.get(block_hash)?
-			.get(transaction_index)?;
+			.get(&transaction_index)?;
 		let receipt = cache.receipts_by_hash.get(receipt_hash)?;
 		Some(receipt.clone())
 	}
@@ -84,7 +84,7 @@ struct ReceiptCache {
 	signed_tx_by_hash: HashMap<H256, TransactionSigned>,
 
 	/// A map of receipt hashes by block hash.
-	transaction_hashes_by_block_and_index: HashMap<H256, HashMap<U256, H256>>,
+	transaction_hashes_by_block_and_index: HashMap<H256, HashMap<usize, H256>>,
 }
 
 impl ReceiptCache {
@@ -93,7 +93,9 @@ impl ReceiptCache {
 		if !receipts.is_empty() {
 			let values = receipts
 				.iter()
-				.map(|(_, receipt)| (receipt.transaction_index, receipt.transaction_hash))
+				.map(|(_, receipt)| {
+					(receipt.transaction_index.as_usize(), receipt.transaction_hash)
+				})
 				.collect::<HashMap<_, _>>();
 
 			self.transaction_hashes_by_block_and_index.insert(*block_hash, values);
